tests/wpt/tests/trusted-types/trusted-types-reporting-clipping-of-sample.tentative.html

<!DOCTYPE html>
<meta charset="UTF-8">
<script src="/resources/testharness.js" ></script>
<script src="/resources/testharnessreport.js"></script>
<script src="./support/csp-violations.js"></script>
<meta http-equiv="Content-Security-Policy" content="trusted-types 'none'">
<meta http-equiv="Content-Security-Policy" content="require-trusted-types-for 'script'">
<meta http-equiv="Content-Security-Policy" content="connect-src 'none'">
<body>
<script>
  const clippedSampleLength = 40;
  const inputAndClippedInput = [
    // Strings with surrogate pairs
    // We use Firefox's behavior as expectations i.e. clip the source as UTF-16
    // string of length 40 ; or of length 41 if the last character is a
    // surrogate pair). See https://github.com/w3c/webappsec-csp/issues/704
    ['ð†'.repeat(clippedSampleLength / 2), 'ð†'.repeat(clippedSampleLength / 2)],
    [`H${'ð‡'.repeat(clippedSampleLength / 2)}`, `H${'ð‡'.repeat(clippedSampleLength / 2)}`],
    [`${'ðˆ'.repeat(clippedSampleLength / 2)}I`, `${'ðˆ'.repeat(clippedSampleLength / 2)}`],
    ['ð‰'.repeat(clippedSampleLength / 2 + 1), 'ð‰'.repeat(clippedSampleLength / 2)],
  ];

  for (const [input, clippedInput] of inputAndClippedInput) {
    promise_test(async t => {
      let violation = await trusted_type_violation_for(TypeError, _ =>
        window.trustedTypes.createPolicy(input, { createHTML: s => s } )
      );
      assert_equals(violation.originalPolicy, "trusted-types 'none'");
      assert_equals(violation.sample, clippedInput);
    }, `Clipping of violation sample for createPolicy(${input})`);

    promise_test(async t => {
      let violation = await trusted_type_violation_for(TypeError, _ =>
        document.createElement("div").innerHTML = input
      );
      assert_equals(violation.originalPolicy, "require-trusted-types-for 'script'");
      assert_equals(violation.sample, `Element innerHTML|${clippedInput}`);
    }, `Clipping of violation sample for Element.innerHTML = "${input}"`);
  }
</script>
