third_party/blink/web_tests/external/wpt/trusted-types/trusted-types-reporting-clipping-of-sample.tentative-expected.txt

This is a testharness.js-based test.
[FAIL] Clipping of violation sample for createPolicy(Hğ‡ğ‡ğ‡ğ‡ğ‡ğ‡ğ‡ğ‡ğ‡ğ‡ğ‡ğ‡ğ‡ğ‡ğ‡ğ‡ğ‡ğ‡ğ‡ğ‡)
  assert_equals: expected "Hğ‡ğ‡ğ‡ğ‡ğ‡ğ‡ğ‡ğ‡ğ‡ğ‡ğ‡ğ‡ğ‡ğ‡ğ‡ğ‡ğ‡ğ‡ğ‡ğ‡" but got "Hğ‡ğ‡ğ‡ğ‡ğ‡ğ‡ğ‡ğ‡ğ‡ğ‡ğ‡ğ‡ğ‡ğ‡ğ‡ğ‡ğ‡ğ‡ğ‡U+d835"
[FAIL] Clipping of violation sample for Element.innerHTML = "Hğ‡ğ‡ğ‡ğ‡ğ‡ğ‡ğ‡ğ‡ğ‡ğ‡ğ‡ğ‡ğ‡ğ‡ğ‡ğ‡ğ‡ğ‡ğ‡ğ‡"
  assert_equals: expected "Element innerHTML|Hğ‡ğ‡ğ‡ğ‡ğ‡ğ‡ğ‡ğ‡ğ‡ğ‡ğ‡ğ‡ğ‡ğ‡ğ‡ğ‡ğ‡ğ‡ğ‡ğ‡" but got "Element innerHTML|Hğ‡ğ‡ğ‡ğ‡ğ‡ğ‡ğ‡ğ‡ğ‡ğ‡ğ‡ğ‡ğ‡ğ‡ğ‡ğ‡ğ‡ğ‡ğ‡U+d835"
Harness: the test ran to completion.
